refactor: PostRepository 인터페이스 분리 및 DI 적용
