+++ src/main/java/org/sopt/repository/PostRepository.java
@@ -1,38 +1,18 @@
 package org.sopt.repository;
 
 import org.sopt.domain.Post;
-import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-@Repository
-public class PostRepository {
+public interface PostRepository {
+    Post save(Post post);
 
-    private final List<Post> postList = new ArrayList<>();
-    private Long nextId = 1L;
+    List<Post> findAll();
 
-    public Post save(Post post) {
-        postList.add(post);
-        return post;
-    }
+    Optional<Post> findById(Long id);
 
-    public List<Post> findAll() {
-        return postList;
-    }
+    boolean deleteById(Long id);
 
-    public Optional<Post> findById(Long id) {
-        return postList.stream()
-                .filter(p -> p.getId().equals(id))
-                .findFirst();
-    }
-
-    public boolean deleteById(Long id) {
-        return postList.removeIf(p -> p.getId().equals(id));
-    }
-
-    public Long generateId() {
-        return nextId++;
-    }
+    Long generateId();
 }
